requirements.txt
fastapi
pydantic
google-generativeai
pillow
numpy
uvicorn
mangum